separated fragments to satisfy CI class requirements
