+++ SAP2000_oM/Enums/BarDesignProcedureType.cs
@@ -25,36 +25,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BH.oM.Structure.Elements;
-using BH.oM.Adapters.SAP2000.Elements;
-using BH.oM.Adapters.SAP2000;
-using BH.Engine.Base;
 
-namespace BH.Engine.Adapters.SAP2000
+namespace BH.oM.Adapters.SAP2000
 {
-    public static partial class Query
+    public enum BarDesignProcedureType : int
     {
-        public static BarAutoMesh BarAutoMesh(this Bar bar)
-        {
-            return bar.FindFragment<BarAutoMesh>();
-        }
+        Steel = 1,
+        Concrete = 2,
+        Aluminum = 7,
+        ColdFormed = 8,
+        NoDesign = 9
 
-        public static BarDesignProcedure BarDesignProcedure(this Bar bar)
-        {
-            return bar.FindFragment<BarDesignProcedure>();
-        }
-
-        public static BarInsertionPointLocation BarInsertionPoint(this Bar bar)
-        {
-            BarInsertionPoint o = bar?.FindFragment<BarInsertionPoint>();
-            return o == null ? BarInsertionPointLocation.Centroid : o.InsertionPoint;
-        }
-
-        public static bool BarModifyStiffnessInsertionPoint(this Bar bar)
-        {
-            BarInsertionPoint o = bar?.FindFragment<BarInsertionPoint>();
-
-            return o == null ? true : o.ModifyStiffness;
-        }
     }
 }
