+++ SAP2000_oM/Fragments/BarDesignProcedure.cs
@@ -25,36 +25,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BH.oM.Structure.Elements;
-using BH.oM.Adapters.SAP2000.Elements;
-using BH.oM.Adapters.SAP2000;
-using BH.Engine.Base;
+using System.ComponentModel;
+using BH.oM.Base;
 
-namespace BH.Engine.Adapters.SAP2000
+namespace BH.oM.Adapters.SAP2000.Elements
 {
-    public static partial class Query
+    public class BarDesignProcedure : IFragment
     {
-        public static BarAutoMesh BarAutoMesh(this Bar bar)
-        {
-            return bar.FindFragment<BarAutoMesh>();
-        }
-
-        public static BarDesignProcedure BarDesignProcedure(this Bar bar)
-        {
-            return bar.FindFragment<BarDesignProcedure>();
-        }
-
-        public static BarInsertionPointLocation BarInsertionPoint(this Bar bar)
-        {
-            BarInsertionPoint o = bar?.FindFragment<BarInsertionPoint>();
-            return o == null ? BarInsertionPointLocation.Centroid : o.InsertionPoint;
-        }
-
-        public static bool BarModifyStiffnessInsertionPoint(this Bar bar)
-        {
-            BarInsertionPoint o = bar?.FindFragment<BarInsertionPoint>();
-
-            return o == null ? true : o.ModifyStiffness;
-        }
+        [Description("Design procedure based on material type.")]
+        public virtual BarDesignProcedureType DesignProcedure { get; set; } = BarDesignProcedureType.NoDesign;
     }
 }
