+++ SAP2000_oM/SAP2000_oM.csproj
@@ -54,9 +54,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Config\SAP2000ActionConfig.cs" />
-    <Compile Include="Enums\Bar.cs" />
+    <Compile Include="Enums\BarDesignProcedureType.cs" />
+    <Compile Include="Enums\BarInsertionPointLocation.cs" />
     <Compile Include="Enums\SDShapeType.cs" />
-    <Compile Include="Fragments\Bar.cs" />
+    <Compile Include="Fragments\BarInsertionPoint.cs" />
+    <Compile Include="Fragments\BarDesignProcedure.cs" />
+    <Compile Include="Fragments\BarAutoMesh.cs" />
     <Compile Include="Fragments\SAP2000Id.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Enums\ShellType.cs" />
